Shuffle chunked builds
Testing this feature on Circle CI, it works great, but I'm noticing that
reasource heavy builds (e.g. CGO builds and the MacOS universal builds)
tend to be configured together and end up in "resource heavy" partitions.

This commit shuffling the builds before the chunk operation, which should
improve this.